+++ plugins/utilities/mood_light.py
@@ -287,7 +287,7 @@ class moodlight(ba.Plugin):
     def on_app_running(self):
         _ba.show_progress_bar()
 
-    def on_plugin_manager_prompt(self):
+    def show_settings_ui(self, source_widget):
         SettingWindow()
 
     def has_settings_ui(self):