+++ plugins/utilities/share_replay.py
@@ -262,6 +262,3 @@ class Loup(ba.Plugin):
 
     def show_settings_ui(self, button):
         SettingWindow()
-
-    def on_plugin_manager_prompt(self):
-        SettingWindow()